Adição do serviço das vendas e categorias, e ajustes nos outros serviços.

+++ cccpharma-server/src/main/java/com/cccpharma/service/SoldProductServiceImpl.java
@@ -6,7 +6,8 @@ import lombok.AllArgsConstructor;
 import lombok.NoArgsConstructor;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
-
+import java.util.ArrayList;
+import java.util.List;
 import static java.util.Objects.isNull;
 
 @Service
@@ -16,6 +17,21 @@ public class SoldProductServiceImpl implements SoldProductService {
 
     @Autowired
     private SoldProductRepository soldProductRepository;
+
+    @Autowired
+    private LotService lotService;
+
+    public List<SoldProduct> findAll() {
+
+        Iterable<SoldProduct> soldProducts = soldProductRepository.findAll();
+        List<SoldProduct> soldProductsResult = new ArrayList<>();
+
+        for(SoldProduct soldProduct : soldProducts) {
+            soldProductsResult.add(soldProduct);
+        }
+
+        return soldProductsResult;
+    }
 
     public SoldProduct findById(Long id) {
 
@@ -31,6 +47,9 @@ public class SoldProductServiceImpl implements SoldProductService {
 
     public SoldProduct save(SoldProduct soldProduct) {
 
+        if(isNull(soldProduct)) {
+            throw new NullPointerException("Sold product is null!");
+        }
         if(isNull(soldProduct.getProduct())) {
             throw new NullPointerException("Product is null!");
         }
@@ -49,6 +68,10 @@ public class SoldProductServiceImpl implements SoldProductService {
         if(isNull(id)) {
             throw new NullPointerException("Sold product id is null!");
         }
+        if(!soldProductRepository.existsById(id)) {
+            throw new RuntimeException("Sold product not found!");
+        }
+
         soldProductRepository.deleteById(id);
     }
